+++ packages/slice-machine/src/components/GenerateSliceWithAiModal/SliceCard.tsx
