+++ packages/slice-machine/src/components/GenerateSliceWithAiModal/index.tsx
